% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketches.R
\name{srht_sketch}
\alias{srht_sketch}
\title{Generate an SRHT sketching matrix}
\usage{
srht_sketch(s, n)
}
\arguments{
\item{s}{Integer number of sampled Hadamard rows.}

\item{n}{Integer input dimension. Must be a power of two.}
}
\value{
A numeric matrix with `s` rows and `n` columns.
}
\description{
Creates a subsampled randomized Hadamard transform (SRHT) sketch. The input
dimension `n` must be a power of two.
}
\examples{
set.seed(1)
S <- srht_sketch(4, 8)
dim(S)

}
